Move cn function to utils
