+++ client/src/components/add-collaborators/AddCollaborators.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useRef, useState } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 import Email from "./Email";
 import { validateEmail } from "./validate-email";
