+++ client/src/components/add-collaborators/Email.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 import { X } from "lucide-react";
 import { validateEmail } from "./validate-email";
