+++ client/src/components/navigation/NavItem.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 import { NavLink } from "react-router";
 
 type NavItemProps = {
