+++ client/src/components/ui/Card.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 import PageSection from "@/components/ui/PageSection";
 
 type CardProps = {
