+++ client/src/components/ui/EditableText.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 import { Pencil } from "lucide-react";
 import { useEffect, useRef, useState } from "react";
