+++ client/src/components/ui/PageSection.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 
 type PageSectionProps = {
   children: React.ReactNode;
