+++ client/src/components/ui/Spinner.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 
 type SpinnerProps = {
   size?: number;
