+++ client/src/components/ui/TextArea.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 
 type TextAreaProps = {
   placeholder?: string;
