+++ client/src/components/ui/calendar.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import { ChevronLeft, ChevronRight } from "lucide-react";
 import { DayPicker } from "react-day-picker";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/cn";
 import { buttonVariants } from "@/components/ui/button";
 
 function Calendar({
